Exception messaging for failure to extend session has been improved.
